+++ routes/Offers.js
@@ -24,9 +24,9 @@ class Offers {
 
         this._program.command('offer', '<id>')
         .description("Retrieve information about an existing offer")
-        .option('--json', 'Write the JSON result to standard stdout')
+        .option('-j, --json', 'Write the JSON result to standard stdout')
         .option('-v, --verbose', 'Use verbose console mode')
-        .on('--help', function(){
+        .on('-h, --help', function(){
             console.log('  Examples:');
             console.log('');
             console.log('    $ rbw offer 57ea7475d78f3ba5aae98935');
@@ -56,7 +56,7 @@ class Offers {
         .option('-n, --name <name>', 'Filter by name')
         .option('--json', 'Write the JSON result to standard stdout')
         .option('-f, --file <filename>', 'Print result to a file in CSV')
-        .on('--help', function(){
+        .on('-h, --help', function(){
             console.log('  Examples:');
             console.log('');
             console.log('    $ rbw offers');
@@ -108,9 +108,9 @@ class Offers {
         });
         this._program.command('catalog', '<id>')
         .description("Retrieve information about an existing catalog")
-        .option('--json', 'Write the JSON result to standard stdout')
+        .option('-j, --json', 'Write the JSON result to standard stdout')
         .option('-v, --verbose', 'Use verbose console mode')
-        .on('--help', function(){
+        .on('-h, --help', function(){
             console.log('  Examples:');
             console.log('');
             console.log('    $ rbw catalog 57ea7475d78f3ba5aae98935');
@@ -138,9 +138,9 @@ class Offers {
         .option('-p, --page <number>', 'Display a specific page')
         .option('-l, --limit <number>', 'Limit to a number of instances per page (max=1000')
         .option('-n, --name <name>', 'Filter by name')
-        .option('--json', 'Write the JSON result to standard stdout')
+        .option('-j, --json', 'Write the JSON result to standard stdout')
         .option('-f, --file <filename>', 'Print result to a file in CSV')
-        .on('--help', function(){
+        .on('-h, --help', function(){
             console.log('  Examples:');
             console.log('');
             console.log('    $ rbw catalogs');
@@ -193,9 +193,9 @@ class Offers {
 
         this._program.command('create catalog', '<name> [description]')
         .description("Create a new catalog")
-        .option('--json', 'Write the JSON result to standard stdout')
+        .option('-j, --json', 'Write the JSON result to standard stdout')
         .option('-v, --verbose', 'Use verbose console mode')
-        .on('--help', function(){
+        .on('-h, --help', function(){
             console.log('  Examples:');
             console.log('');
             console.log('    $ rbw create catalog "aCatalog"');
@@ -223,7 +223,7 @@ class Offers {
         .option('-l, --list <items>', 'A list', function(val) {
             return val.split(',');})
         .option('-v, --verbose', 'Use verbose console mode')
-        .on('--help', function(){
+        .on('-h, --help', function(){
             console.log('  Examples:');
             console.log('');
             console.log('    $ rbw delete catalog 57ea7475d78f3ba5aae98935');