+++ routes/Status.js
@@ -23,8 +23,17 @@ class Status {
 
         this._program.command('status api')
         .description("Display status of API")
-        .option('--json', 'Write the JSON result to standard stdout')
+        .option('-j, --json', 'Write the JSON result to standard stdout')
         .option('-v, --verbose', 'Use verbose console mode')
+        .on('-h, --help', function(){
+            console.log('  Examples:');
+            console.log('');
+            console.log('    $ rbw status api');
+            console.log('');
+            console.log('  Details:');
+            console.log('');
+            console.log('    The option `--json` exports the JSON object representing the user to the console');
+        })
         .action(function (commands) {
 
             var options= {